requirements.txt
# MCP Server
mcp>=1.0.0

# Vector Store and Embeddings
chromadb>=0.4.0
openai>=1.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0
